backend_app/requirements.txt
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pillow==8.0.1
psycopg2==2.8.6
